Move `Guard` and `Accumulator` implementations into `verifier` submodule.
